requirements.txt
usd-core=24.11
pillow=11.0.0
numpy=2.2.0